backend/parser/requirements.txt
beautifulsoup4==4.9.3
psycopg2-binary==2.9.1
python-dotenv==0.19.0
requests==2.25.1
schedule==1.1.0
